Pass current value into getter function
According to the docs you should have access to the current value and setter when creating a `to` function. However this file currently does'nt not pass any arguments to the `valueOrGetter`. Now the `valueOrGetter` is called with the current value if its a function. We should instead return the new value or throw.